requirements.txt
instagrapi
python-dotenv 